requirements.txt
pure-python-adb
Pillow

numpy
tensorflow
scipy

notebook
matplotlib

multipledispatch
pytest


opencv-python
opencv-contrib-python


scikit-image
pandas
sklearn
